Overhaul of the rendering core. Split a lot out into partial files for SDL3Renderer as it is unsurprisingly huge. Added a lot of tests and expanded on the headless renderer to be an actual thing, not just a no-op.

+++ src/Brine2D/Systems/Rendering/SpriteComponent.cs
@@ -57,5 +57,5 @@ public class SpriteComponent : Component
     /// Rendering layer/order (higher = drawn on top).
     /// Used by the batching system to sort sprites.
     /// </summary>
-    public int Layer { get; set; } = 0;
+    public byte Layer { get; set; } = 0;
 }
